edit tampilan menu print rekap

+++ application/controllers/Auth.php
@@ -41,8 +41,11 @@ class Auth extends CI_Controller {
 	}
 
 	public function logout(){
+		if (!empty($this->session->file)) {
 		$un = $this->session->file;
-		unlink($_SERVER['DOCUMENT_ROOT'].'/sensasiq/assets/qrimg/'.$un); 
+		unlink($_SERVER['DOCUMENT_ROOT'].'/sensasiq/assets/qrimg/'.$un);
+		}
+		
 		$this->session->sess_destroy();
 		$this->session->set_flashdata('message', 'Anda telah keluar.');
 		redirect('auth');
